move getMessage from pages/popup to helpers
(as can be used anywhere, not just popup)
